+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `SET_PRODUCT_QUERY` action
 
 ## [0.2.0] - 2019-08-27
 ### Added
@@ -13,7 +15,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `selectedItem` and `selectedQuantity` to context state
 
 ## [0.1.0] - 2019-07-02
-
 ### Added
-
 - Initial release.